계산기 애플리케이션의 반환 타입 및 DTO 변경
계산기 애플리케이션의 반환 타입을 변경하고, 이에 따른 DTO의 수정이 이루어졌습니다.

주요 변경사항:
1. Calculator 클래스의 calculate 메소드의 반환 타입이 Calculation에서 int로 변경되었습니다.
2. 이에 따라 PostCalculationRequest 클래스에서 반환하는 CalculationResponseDto의 생성자 인자가 수정되었습니다.
3. CalculationResponseDto와 CalculationListRequestDto에서 불필요한 필드를 제거하고, CalculationListResponseDto를 새로 생성하여 필요한 필드만을 가지도록 하였습니다.
4. CalculationRepository의 findAll 메소드에서 calculations 리스트를 출력하는 로그를 추가하였습니다.
